Treat warnings as errors; fix what that surfaced; bump actions/checkout
Wire -Wall -Wextra -Werror (GCC/Clang) and /W4 /WX (MSVC) into our own
targets only (libs, the CLI, the test executables) so a warning like
the earlier unused-variable one fails CI instead of slipping through.
-Wmissing-field-initializers is excluded: this code relies on C++20
designated initializers to skip fields that already default sanely
(std::optional members etc), which is exactly what that warning flags.

Fixes found by enabling it:
- DeviceProfile.h: to_string(SonyModel) had no case for WF1000XM6,
  silently falling back to "Unknown" for that model - a real bug, not
  just a warning.
- BluetoothWrapper.cpp: recv() can return a negative count on error;
  clamp it before using it as a size_t bound instead of letting a
  negative value silently promote to a huge unsigned one.
- ByteMagic.cpp: signed/unsigned loop comparison.
- IDeviceService.h, DeviceEventDispatcherTests.cpp,
  ReliabilityTests.cpp: unused parameter/variable/return value.

Also bumps actions/checkout to v5 (Node 24) - v4 is being forced onto
a runtime it wasn't built for and will stop working per GitHub's
Node 20 deprecation notice.

+++ libs/sony-transport/legacy/ByteMagic.cpp
@@ -48,7 +48,7 @@ std::string MACBytesToString(unsigned char addr[])
 {
 	// Create hex string from the bytes.
 	std::stringstream mac_addr_str;
-	for (int i = 0; i < MAC_ADDR_BYTES_LEN; i++)
+	for (size_t i = 0; i < MAC_ADDR_BYTES_LEN; i++)
 	{
 		mac_addr_str << std::hex << (((addr[MAC_ADDR_BYTES_LEN - 1 - i] & 0xF0) >> 4) & 0xF);
 		mac_addr_str << std::hex << (addr[MAC_ADDR_BYTES_LEN - 1 - i] & 0x0F);
